pyproject.toml
[tool.poetry]
name = "solar-backend-fastapi"
version = "0.1.0"
description = ""
authors = ["ilevk <kcs93023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^3.7.1"
openai = "1.2.0"
fastapi = "^0.111.1"
uvloop = "^0.19.0"
pydantic-settings = "^2.3.4"
orjson = "^3.10.6"
retry = "^0.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
